+++ metabolograf/CurrentValues.cpp
@@ -60,7 +60,7 @@ void CurrentValues::Init()
 //------------------------------------------------------
 void CurrentValues::SetSelected(int new_selected)
 {
-	if (new_selected>=database->getCount()) 
+	if (new_selected>= static_cast<int>(database->getCount()))
 		selected = -1;
 	else if (selected == new_selected) 
 		return;
@@ -86,7 +86,7 @@ void CurrentValues::OnPaint()
 	//ugc.SetDrawColor(100, 100, 100);
 	//ugc.FillRectangle(0, 0, Width, 20 * dpiX);
 
-	ugc.SetTextSize(10*dpiX);
+	ugc.SetTextSize(static_cast<int>(10*dpiX));
 	ugc.SetAlign(Align::CENTER);
 
 	//ugc.SetDrawColor(255, 255, 255);
@@ -100,14 +100,14 @@ void CurrentValues::OnPaint()
 	int startY = 0;// space * 2;
 
 	//ugc.DrawLine(w / 2, startY, w / 2, variable_names.size() / 2 * space * 3+startY);
-	const int border = 2 * dpiX;
-	for (int i = 0; i < variable_names.size(); i++)
+	const int border = static_cast<int>(2 * dpiX);
+	for (size_t i = 0; i < variable_names.size(); i++)
 	{
 		string& vname = variable_names[i];
 		if (vname=="ЧСС" && !database->isVariableExists(vname)) continue;//если чсс не учитывается и нет такой переменной то рисовать не надо
 		
-		int yi = i / 2;
-		yi *= (space * 2.7);
+		int yi = static_cast<int>(i / 2);
+		yi *= static_cast<int>(space * 2.7);
 		yi += startY+border;
 
 		int xi = 0;
@@ -118,13 +118,13 @@ void CurrentValues::OnPaint()
 			//	ugc.DrawLine(0, yi-space/2, Width, yi-space/2);
 		}
 		ugc.SetDrawColor(60, 60, 60);
-		ugc.FillRectangle(xi+2*dpiX, yi, w/2-4*dpiX, space*2.5);
+		ugc.FillRectangle(static_cast<int>(xi+2*dpiX), yi, static_cast<int>(w/2-4*dpiX), static_cast<int>(space*2.5));
 		ugc.SetDrawColor(255, 255, 255);
-		ugc.FillRectangle(xi + 2 * dpiX + border, yi+space, w / 2 - 4 * dpiX - 2*border, space*1.5-border);
+		ugc.FillRectangle(static_cast<int>(xi + 2 * dpiX + border), yi+space, static_cast<int>(w / 2 - 4 * dpiX - 2*border), static_cast<int>(space*1.5-border));
 		ugc.SetDrawColor(225, 225, 225);
-		ugc.SetTextSize(9 * dpiX);
-		ugc.DrawString(lables[i], xi+w/4, yi+2*dpiX);
-		ugc.SetTextSize(22*dpiX);
+		ugc.SetTextSize(static_cast<int>(9 * dpiX));
+		ugc.DrawString(lables[i], xi+w/4, static_cast<int>(yi+2*dpiX));
+		ugc.SetTextSize(static_cast<int>(22*dpiX));
 
 		
 		double value = 0;
@@ -135,26 +135,26 @@ void CurrentValues::OnPaint()
 		ugc.SetDrawColor(0, 0, 0);
 
 		ugc.SetBold(true);
-		ugc.DrawString(ToString(vname,value), xi+w/4, yi+space*0.6);
+		ugc.DrawString(ToString(vname,value), xi+w/4, static_cast<int>(yi+space*0.6));
 		ugc.SetBold(false);
 	}
 
 	
 	
 	int yi = (variable_names.size()) / 2;
-	yi *= (space * 2.7);
+	yi *= static_cast<int>(space * 2.7);
 	yi += startY+border;
 
 	int xi = w / 2;
 	
 	ugc.SetDrawColor(60, 60, 60);
-	ugc.FillRectangle(xi + 2 * dpiX, yi, w / 2 - 4 * dpiX, space*2.5);
+	ugc.FillRectangle(static_cast<int>(xi + 2 * dpiX), yi, static_cast<int>(w / 2 - 4 * dpiX), static_cast<int>(space*2.5));
 	ugc.SetDrawColor(255, 255, 255);
-	ugc.FillRectangle(xi + 2 * dpiX + border, yi + space, w / 2 - 4 * dpiX - 2 * border, space*1.5 - border);
+	ugc.FillRectangle(static_cast<int>(xi + 2 * dpiX + border), yi + space, static_cast<int>(w / 2 - 4 * dpiX - 2 * border), static_cast<int>(space*1.5 - border));
 	ugc.SetDrawColor(225, 225, 225);
-	ugc.SetTextSize(9 * dpiX);
-	ugc.DrawString("Вдох-Выдох (%)", xi + w / 4, yi + 2 * dpiX);
-	ugc.SetTextSize(22 * dpiX);
+	ugc.SetTextSize(static_cast<int>(9 * dpiX));
+	ugc.DrawString("Вдох-Выдох (%)", xi + w / 4, static_cast<int>(yi + 2 * dpiX));
+	ugc.SetTextSize(static_cast<int>(22 * dpiX));
 
 
 	double value = 0;
@@ -167,10 +167,10 @@ void CurrentValues::OnPaint()
 	ugc.SetBold(true);
 	//ugc.DrawString(DoubleToString(CalculateErr()), xi + w / 4, yi + space*0.6);
 	//ugc.SetTextSize(18 * dpiX);
-	ugc.DrawString(DoubleToString(CalculateErr()), xi + w / 8, yi + space*0.6);
+	ugc.DrawString(DoubleToString(CalculateErr()), xi + w / 8, static_cast<int>(yi + space*0.6));
 	ugc.SetDrawColor(140, 140, 140);
 	
-	ugc.DrawString(DoubleToString(OldCalculateErr()), xi + w * 3 / 8, yi + space*0.6);
+	ugc.DrawString(DoubleToString(OldCalculateErr()), xi + w * 3 / 8, static_cast<int>(yi + space*0.6));
 	ugc.SetBold(false);
 	ugc.SetAlign(Align::LEFT);
 }
@@ -194,7 +194,7 @@ double CurrentValues::CalculateErr()
 	if (database->getCount() == 0) return 0;
 	double result = 0;
 	int count = 0;
-	for (int i = 0; i < database->getCount(); ++i)
+	for (size_t i = 0; i < database->getCount(); ++i)
 	{
 		if (database->getChecked(i))
 		{
