+++ metabolograf/VolumeErrorDialog.cpp
@@ -43,7 +43,7 @@ BOOL VolumeErrorDialog::OnInitDialog()
 	CRect rect;
 	GetClientRect(&rect);
 	DPIX dpix;
-	rect.bottom -= 30 * (double)dpix;
+	rect.bottom -= static_cast<LONG>(30 * (double)dpix);
 	hist.Create(NULL, NULL, WS_VISIBLE | WS_CHILD, rect, this, IDC_BIG_PLOT);
 	hist.SetBounds();
